+++ java/src/main/java/io/pingui/CliRunMode.java
@@ -4,6 +4,7 @@ package io.pingui;
 public enum CliRunMode {
     GUI,
     EXPORT,
+    TELEMETRY_RETENTION,
     DAEMON,
     STOP,
     STATUS